Fix missing telemetry opt out toast (#23570)
* Fix telemetry opt out toast

* fix imports

* cleanup
