+++ src/sql/workbench/contrib/welcome/telemetryOptOut/electron-sandbox/telemetryOptOut.ts
@@ -7,14 +7,10 @@ import { IStorageService } from 'vs/platform/storage/common/storage';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
 import { INotificationService } from 'vs/platform/notification/common/notification';
-import { IExperimentService } from 'vs/workbench/contrib/experiments/common/experimentService';
-import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { IExtensionGalleryService } from 'vs/platform/extensionManagement/common/extensionManagement';
 import { IProductService } from 'vs/platform/product/common/productService';
 import { IHostService } from 'vs/workbench/services/host/browser/host';
-import { AbstractTelemetryOptOut } from 'vs/workbench/contrib/welcome/telemetryOptOut/browser/telemetryOptOut';
+import { AbstractTelemetryOptOut } from 'sql/workbench/contrib/welcome/telemetryOptOut/browser/telemetryOptOut';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
-import { IJSONEditingService } from 'vs/workbench/services/configuration/common/jsonEditing';
 import { INativeHostService } from 'vs/platform/native/common/native';
 
 export class NativeTelemetryOptOut extends AbstractTelemetryOptOut {
@@ -25,20 +21,16 @@ export class NativeTelemetryOptOut extends AbstractTelemetryOptOut {
 		@INotificationService notificationService: INotificationService,
 		@IHostService hostService: IHostService,
 		@ITelemetryService telemetryService: ITelemetryService,
-		@IExperimentService experimentService: IExperimentService,
-		@IConfigurationService configurationService: IConfigurationService,
-		@IExtensionGalleryService galleryService: IExtensionGalleryService,
 		@IProductService productService: IProductService,
 		@IEnvironmentService environmentService: IEnvironmentService,
-		@IJSONEditingService jsonEditingService: IJSONEditingService,
 		@INativeHostService private readonly nativeHostService: INativeHostService
 	) {
-		super(storageService, openerService, notificationService, hostService, telemetryService, experimentService, configurationService, galleryService, productService, environmentService, jsonEditingService);
+		super(storageService, openerService, notificationService, hostService, telemetryService, productService, environmentService);
 
 		this.handleTelemetryOptOut();
 	}
 
 	protected getWindowCount(): Promise<number> {
-		return this.nativeHostService ? this.nativeHostService.getWindowCount() : Promise.resolve(0); // {{SQL CARBON EDIT}} Tests run without UI context so electronService is undefined in that case
+		return this.nativeHostService ? this.nativeHostService.getWindowCount() : Promise.resolve(0);
 	}
 }
